+++ src/commands/sync.rs
@@ -14,13 +14,14 @@ use rayon::prelude::*;
 use serde_json::to_string_pretty;
 use ureq::Agent;
 use semver::{Error, Version};
+use tokio::io::AsyncWriteExt;
 use tracing::{debug, error, info, warn};
-use crate::aliases::{ModFileName, ModID, ModVersion};
+use crate::aliases::{ModFileName, ModID, ModIDInt, ModName, ModVersion};
 use crate::rustique_errors::RustiqueError;
 use crate::api::api_structs::{Mod, ModInfo, Releases};
-use crate::utils::{RustiqueOptions, get_current_time, extract_all_mods_metadata, elapsed_footer, notice};
+use crate::utils::{RustiqueOptions, get_current_time, extract_all_mods_metadata, elapsed_footer, notice, is_today, get_expanded_path, timestamp_older_than};
 use crate::api::client::{ApiClient, ModApiFetch};
-use crate::config_manager::get_config;
+use crate::config_manager::{get_config, CONFIG_DEFAULT_DIR};
 use crate::install_manager::Install;
 use crate::rustique_errors::RustiqueError::UrlParseError;
 use crate::version_management::{parse_latest_version, parse_version};
@@ -30,16 +31,40 @@ pub struct RustiqueSyncJson {
     #[serde(rename = "RustiqueSync")]
     pub rustique_sync: HashMap<ModID, ModSyncInfo>,
     pub last_sync: String,
+    pub last_modid_sync: String,
 }
 
 impl RustiqueSyncJson {
     pub fn new() -> RustiqueSyncJson {
         Self {
             rustique_sync: HashMap::<ModID, ModSyncInfo>::new(),
+            last_sync: get_current_time(),
+            last_modid_sync: get_current_time(),
+        }
+    }
+}
+
+#[derive(Deserialize, Serialize, Debug)]
+pub struct ModIDSync {
+    pub all_mods: HashMap<ModName, ModIDSyncData>,
+    pub last_sync: String,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct ModIDSyncData {
+    pub mod_id: ModID,
+    pub modid_strs: Vec<String>
+}
+impl ModIDSync {
+    pub fn new() -> ModIDSync {
+        Self {
+            all_mods: HashMap::new(),
             last_sync: String::new(),
         }
     }
 }
+
+
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct ModSyncInfo {
@@ -62,22 +87,29 @@ pub async fn handle_sync_call(mod_dir: &PathBuf) {
 }
 
 pub const SYNC_FILE_NAME: &str = "rustique-sync.json";
+pub const MODID_SYNC_FILE_NAME: &str = "mod-id-sync.json";
 
-pub fn parse_sync_file(mod_dir: &PathBuf) -> Result<RustiqueSyncJson, RustiqueError> {
-    let mut file = File::open(mod_dir.join(SYNC_FILE_NAME)).map_err(|e| RustiqueError::IoError {
-        context: format!("Unable to open {}", SYNC_FILE_NAME),
+
+pub fn parse_json_file<T>(file_path: &PathBuf) -> Result<T, RustiqueError>
+where
+    T: for<'de> serde::Deserialize<'de>
+{
+    let filename = file_path.file_name().unwrap().to_string_lossy().to_string();
+
+    let mut file = File::open(file_path).map_err(|e| RustiqueError::IoError {
+        context: format!("Unable to open {}", filename),
         source: e,
     })?;
 
     let mut file_contents = String::new();
     file.read_to_string(&mut file_contents).map_err(|e| RustiqueError::IoError {
-        context: format!("Failure while reading from file {}", SYNC_FILE_NAME),
+        context: format!("Failure while reading from file {}", filename),
         source: e
     })?;
 
-    let json = serde_json5::from_str::<RustiqueSyncJson>(&file_contents)
+    let json = serde_json5::from_str::<T>(&file_contents)
         .map_err(|e| RustiqueError::JsonError {
-            context: format!("Json parsing Error for {}", SYNC_FILE_NAME),
+            context: format!("Json parsing Error for {}", filename),
             source: e
         })?;
 
@@ -88,22 +120,54 @@ pub fn parse_sync_file(mod_dir: &PathBuf) -> Result<RustiqueSyncJson, RustiqueEr
 
 
 pub async fn sync(mod_dir: &PathBuf) -> Result<(), RustiqueError> {
-    notice("Syncing...", Option::from(comfy_table::Color::Green), vec![Attribute::Bold]);
+
     let start_time = Instant::now();
     let config = get_config().read().unwrap();
+    mod_id_sync(false).await?;
+
+    notice("Syncing...", Option::from(comfy_table::Color::Blue), vec![Attribute::Bold]);
 
     // check if rustique-sync.json exists
     // if so, parse the file for updating
     // if not, do all the sync process and then write a new file
-    let file_path = mod_dir.join(SYNC_FILE_NAME);
-    debug!("sync file: {}", file_path.display());
+    let sync_file_path = mod_dir.join(SYNC_FILE_NAME);
+    debug!("sync file: {}", sync_file_path.display());
 
-    let mut sync_data = RustiqueSyncJson {
-        rustique_sync: HashMap::new(),
-        last_sync: get_current_time(),
+    let mut sync_data = if sync_file_path.exists() {
+        match parse_json_file::<RustiqueSyncJson>(&sync_file_path) {
+            Ok(json) => json,
+            Err(e) => {
+                // delete the sync file because the json changed
+                tokio::fs::remove_file(&sync_file_path).await?;
+
+                // return a blank slate to keep going
+                RustiqueSyncJson::new()
+            }
+        }
+    } else {
+       RustiqueSyncJson::new()
+    };
+
+    if timestamp_older_than(24, &sync_data.last_modid_sync) {
+        // update the database
+        mod_id_sync(true).await?;
+        sync_data.last_modid_sync = get_current_time();
+    }
+
+    let mod_id_fp = get_expanded_path(PathBuf::from(CONFIG_DEFAULT_DIR).join(MODID_SYNC_FILE_NAME));
+    let id_sync_data = match parse_json_file::<ModIDSync>(&mod_id_fp) {
+        Ok(json) => json,
+        Err(e) => {
+            // this shouldn't fail, but if it does, rerun the mod_id_sync()
+            info!("mod id sync Error: {}", e.to_string());
+            mod_id_sync(true).await?
+        }
     };
 
     let installed_mods = extract_all_mods_metadata(mod_dir)?;
+
+    // clean sync data first so latest info takes priority
+    sync_data.rustique_sync.clear();
 
     installed_mods.iter().for_each(|(mod_filename, mod_info)| {
         let version = if let Ok(parsed_version) = parse_version(mod_info.version.clone().unwrap_or_default()) {
@@ -115,9 +179,31 @@ pub async fn sync(mod_dir: &PathBuf) -> Result<(), RustiqueError> {
 
         info!("VERSION Parsed: {} for {}", version, mod_info.mod_id);
 
+        // check here for bad mod_id
+        let mod_id = if mod_info.mod_id.is_empty() {
+            info!("{} has an empty mod id, attempting find it...", mod_filename);
+            match id_sync_data.all_mods.get(&mod_info.name) {
+                Some(id) => id.mod_id.clone(),
+                None => {
+                    match id_sync_data.all_mods.values().find(|mod_data| mod_data.modid_strs.contains(&mod_info.name)) {
+
+                        Some(id) => id.mod_id.clone(),
+                        None => {
+                            warn!("Unable to determine the mod_id for {} - {}.\n\r\t Their modinfo.json is malformed and no information provided allowed Rustique to determine it.\n\r\t \
+                             Please contact the author to correct their modinfo.json file", mod_info.name.bright_red().bold(), mod_filename.bright_red().bold());
+                            // We let the value be blank, rustique will report the issue again, but continue
+                            "".to_string()
+                        }
+                    }
+                }
+            }
+        } else {
+            mod_info.mod_id.clone()
+        };
+
         sync_data
             .rustique_sync
-            .entry(mod_info.mod_id.clone())
+            .entry(mod_id)
             .or_insert_with(|| ModSyncInfo {
                 installed_version: version.clone(),
                 file_name: mod_filename.clone(),
@@ -127,12 +213,15 @@ pub async fn sync(mod_dir: &PathBuf) -> Result<(), RustiqueError> {
             });
     });
 
+    let im = installed_mods.keys().clone().collect::<Vec<&String>>();
+    info!("Installed mods: {:?}", im);
+
     // Create API client and fetch mods in parallel using async
     let client = ApiClient::new();
     let result: HashMap<ModID, Mod> = client
         .fetch_mods_parallel(
-            installed_mods.into_values().map(|m|m.mod_id).collect())
-        .await?;
+            sync_data.rustique_sync.keys().cloned().collect()
+        ).await?;
 
     result.iter().for_each(|(mod_id, mod_info): (&ModID, &Mod)| {
         let (mod_version, download_url) = parse_latest_version(&mod_info.mod_json.releases);
@@ -155,20 +244,17 @@ pub async fn sync(mod_dir: &PathBuf) -> Result<(), RustiqueError> {
 
     // Write the sync data to file
     let data = sync_data;
-    let json = to_string_pretty(&data).map_err(|e| RustiqueError::JsonError {
-        context: "Failure while making the sync json pretty".to_string(),
-        source: serde_json5::Error::from(std::io::Error::new(std::io::ErrorKind::Other, e)),
-    })?;
+    let json = prettify(&data, "Sync")?;
 
     // Use tokio's async file operations
-    let mut file = tokio::fs::File::create(file_path)
+    let mut file = tokio::fs::File::create(sync_file_path)
         .await
         .map_err(|e| RustiqueError::IoError {
             context: format!("Error writing sync file to mod_dir: {}", mod_dir.to_string_lossy()),
             source: e,
         })?;
 
-    tokio::io::AsyncWriteExt::write_all(&mut file, json.as_bytes())
+    AsyncWriteExt::write_all(&mut file, json.as_bytes())
         .await?;
         // .map_err(|e| RustiqueError::ApiError {
         //     context: format!("Error writing data to sync file: {}", file_path.to_string_lossy()),
@@ -180,4 +266,81 @@ pub async fn sync(mod_dir: &PathBuf) -> Result<(), RustiqueError> {
     }
 
     Ok(())
+}
+
+pub async fn mod_id_sync(force: bool) -> Result<ModIDSync, RustiqueError> {
+
+
+    let config = get_config().read().unwrap();
+    let start_time = Instant::now();
+
+    let config_dir = get_expanded_path(PathBuf::from(CONFIG_DEFAULT_DIR));
+    let file_path = config_dir.join(MODID_SYNC_FILE_NAME);
+
+    let mut file_data = if file_path.exists() {
+        match parse_json_file::<ModIDSync>(&file_path) {
+            Ok(json) => json,
+            Err(e) => {
+                // delete the file and try again
+                info!("mod_id_sync json parse error: {}", e);
+                tokio::fs::remove_file(&file_path).await?;
+                println!("mod_id_sync json parse error: {}", e);
+                ModIDSync::new()
+            }
+        }
+    } else {
+        ModIDSync {
+            all_mods: HashMap::new(),
+            last_sync: get_current_time(),
+        }
+    };
+
+    //
+    if file_data.all_mods.is_empty() || force || timestamp_older_than(24, &file_data.last_sync){
+
+        notice("Daily ModID Sync...", Some(comfy_table::Color::Blue), vec![Attribute::Bold]);
+
+        let client = ApiClient::new();
+        // get all mod info
+        let all_mods = client.fetch_all_mods().await?;
+
+        file_data.all_mods = all_mods.mods
+            .into_iter()
+            .filter_map(|m| {
+                Some((m.name.unwrap_or_default(), ModIDSyncData {
+                    mod_id: m.mod_id.to_string(),
+                    modid_strs: m.mod_id_strs,
+                }))
+            })
+            .collect();
+
+        debug!("file_data {:?}", file_data);
+
+        info!("Attempting to write ModID Sync to {}", file_path.display());
+
+        let json = prettify(&file_data, "Sync ModID")?;
+        let mut open_file = tokio::fs::File::create(file_path).await.map_err(|e| RustiqueError::IoError {
+            context: format!("Error writing sync file to config dir: {}", config_dir.to_string_lossy()),
+            source: e,
+        })?;;
+        AsyncWriteExt::write_all(&mut open_file, json.as_bytes()).await?;
+
+        info!("ModID Sync file written successfully");
+    }
+
+    if config.show_execution_time && force {
+        elapsed_footer(start_time, "ModID Sync");
+    }
+
+    Ok(file_data)
+}
+
+pub fn prettify<T>(data: T, command_type: &str) -> Result<String, RustiqueError>
+    where
+    T: serde::Serialize {
+
+    to_string_pretty(&data).map_err(|e| RustiqueError::JsonError {
+            context: format!("Failure while making the {} json pretty", command_type),
+            source: serde_json5::Error::from(std::io::Error::new(std::io::ErrorKind::Other, e)),
+        })
 }